Create a base test class for user mgmt tests
