+++ services/common/test_utils.py
@@ -170,98 +170,6 @@ class BaseOfficeServiceIntegrationTest(BaseIntegrationTest):
         self.office_redis_patcher.stop()
 
 
-class BaseUserManagementIntegrationTest(BaseIntegrationTest):
-    """Base class for User Management integration tests."""
-
-    def setup_method(self):
-        """Set up User Management specific test environment."""
-        # Use selective HTTP patches that don't interfere with TestClient
-        self.http_patches = [
-            # Patch async httpx client (most likely to be used for real external calls)
-            patch(
-                "httpx.AsyncClient._send_single_request",
-                side_effect=AssertionError(
-                    "Real HTTP call detected! AsyncClient._send_single_request was called"
-                ),
-            ),
-            # Patch requests (commonly used for external calls)
-            patch(
-                "requests.adapters.HTTPAdapter.send",
-                side_effect=AssertionError(
-                    "Real HTTP call detected! requests HTTPAdapter.send was called"
-                ),
-            ),
-            # Patch urllib (basic HTTP library)
-            patch(
-                "urllib.request.urlopen",
-                side_effect=AssertionError(
-                    "Real HTTP call detected! urllib.request.urlopen was called"
-                ),
-            ),
-            # Note: We don't patch httpx.Client.send because TestClient uses it internally
-        ]
-
-        # Start all HTTP detection patches
-        for http_patch in self.http_patches:
-            http_patch.start()
-
-        # User Management specific environment
-        os.environ["DB_URL_USER_MANAGEMENT"] = "sqlite:///:memory:"
-        os.environ["TOKEN_ENCRYPTION_SALT"] = "dGVzdC1zYWx0LTE2Ynl0ZQ=="
-
-        # Reload the database module to pick up the new environment variable
-        import importlib
-
-        import services.user_management.database
-
-        importlib.reload(services.user_management.database)
-
-        # Actually create the database tables for in-memory database
-        import asyncio
-
-        from services.user_management.database import create_all_tables
-
-        asyncio.run(create_all_tables())
-
-        # Import and create test client
-        from services.user_management.main import app
-
-        self.app = app
-        self.client = self.create_test_client(app)
-
-        # Set up authentication overrides
-        self._override_auth()
-
-    def teardown_method(self):
-        """Clean up User Management specific patches."""
-        # Stop all patches
-        for http_patch in self.http_patches:
-            http_patch.stop()
-
-        self.app.dependency_overrides.clear()
-        if hasattr(self, "_patcher"):
-            self._patcher.stop()
-
-    def _override_auth(self):
-        """Override authentication for testing."""
-        from services.user_management.auth.clerk import get_current_user
-
-        async def mock_get_current_user():
-            return "user_123"
-
-        async def mock_verify_user_ownership(
-            current_user_id: str, resource_user_id: str
-        ):
-            return None
-
-        self.app.dependency_overrides[get_current_user] = mock_get_current_user
-        patcher = patch(
-            "services.user_management.auth.clerk.verify_user_ownership",
-            side_effect=mock_verify_user_ownership,
-        )
-        self._patcher = patcher.start()
-
-
 # For services that don't need specific setup, they can use the selective HTTP patches
 class BaseSelectiveHTTPIntegrationTest(BaseIntegrationTest):
     """Base class for integration tests that need to allow TestClient but block external HTTP calls."""
